requirements.txt
Django>=5.1.2
channels>=4.0.0
asgiref>=3.7.0
pytz>=2023.3
sqlparse>=0.4.4
django-widget-tweaks>=1.5.0
Pillow>=10.0.0
xhtml2pdf>=0.2.11
reportlab>=4.0.0
PyPDF2>=3.0.0
django-ipware>=5.0.0
twilio>=8.0.0
requests>=2.31.0
django-phonenumber-field[phonenumberslite]>=8.0.0
